fix(14): GPU dtype mismatch in jax_session and missing sys.path in valid03
jax.random.categorical returns int64 on GPU but int32 on CPU, causing
lax.scan carry dtype mismatch. Cast choice to int32 explicitly.

valid03_cross_platform.py only added src/ to sys.path but config.py
lives at the project root — add the project root as well.

+++ src/prl_hgf/simulation/jax_session.py
@@ -229,7 +229,7 @@ def _run_session(
         # prev_choice == -1 (sentinel) → stick is all-False → zero stickiness
         stick = (prev_choice == jnp.arange(3)).astype(jnp.float32)
         logits = beta * p_reward + zeta * stick
-        choice = jax.random.categorical(choice_key, logits)
+        choice = jax.random.categorical(choice_key, logits).astype(jnp.int32)
 
         # Step 4: Bernoulli reward sampling for the chosen cue
         reward = jax.random.bernoulli(reward_key, cue_probs[choice])
